requirements.txt
Django==5.2.8
python-decouple==3.8
django-taggit==6.1.0
whitenoise==6.9.0
gunicorn==23.0.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
